Implementation of a UI to add user-configured hotkeys. Users can now add and change hotkeys to whatever they like more or less. It's also easy to add new commands - simply modify the PlayerCommands class and add a new static property, and make sure to include this in the AllCommands method as well. Adding volume control in this manner should be simple.
I added some functionality to MapConfig to enable the saving of the arbitrary list to a config file - it is now possible to have mapconfigentries that store a dictionary of strings rather than a plain string. NB: None of this has been stress tested!

Also worth mentioning that I used a 3rd-party provided class that implements an ObservableDictionary so that I could bind it to the hotkeys grid and still have it notify of property changes. The code seems to be clean enough, but keep an eye on it for problems.

git-svn-id: http://elpis-pandora-client.googlecode.com/svn/trunk@94 03ac0622-d976-f640-46ab-04ada77f7b90